Django/accounting/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.2
